feat: add download button in app terminal
